pyproject.toml
[tool.poetry]
name = "accreditation-api"
version = "0.1.0"
description = "The backend API for MIT Bengaluru's Accreditation System."
authors = ["Alphaspiderman <68412330+Alphaspiderman@users.noreply.github.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
sanic = {extras = ["ext"], version = "^24.6.0"}
sanic-ext = "^23.12.0"
mayim = {extras = ["postgres"], version = "^1.1.0"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
taskipy = "^1.14.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
aiohttp = "^3.10.10"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
pre-commit = "^4.0.1"

[tool.taskipy.tasks]
server = { cmd = "python -m api.server", help = "Runs the API Server"}
lint = { cmd = "pre-commit run", help = "Lints project" }
lintall = { cmd = "pre-commit run --all-files", help = "Lints project" }
precommit = { cmd = "pre-commit install", help = "Installs the pre-commit git hook" }
format = { cmd = "black api", help = "Runs the black python formatter" }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
